Refactor Util
