+++ backend/src/Database/RunnerDataLineSaver.php
@@ -2,7 +2,7 @@
 
 namespace App\Database;
 
-use App\Misc\Util;
+use App\Misc\Util\CommonUtil;
 use App\Model\RunnersCsvDataLine\DataLine;
 
 class RunnerDataLineSaver
@@ -25,7 +25,7 @@ class RunnerDataLineSaver
         $lastname = $dataLine->getRunnerLastname();
         $gender = $dataLine->getRunnerGender();
         $birthYear = $dataLine->getRunnerBirthYear();
-        $category = Util::getFfaCategoryFromBirthYear($birthYear);
+        $category = CommonUtil::getFfaCategoryFromBirthYear($birthYear);
 
         $this->getQuery()->bindParam(':id', $id);
         $this->getQuery()->bindParam(':isTeam', $isTeam);
