+++ backend/src/Responder/CategoriesResponder.php
@@ -4,11 +4,10 @@
 namespace App\Responder;
 
 
-use App\Database\DAO;
 use App\Database\Entity\Runner;
 use App\Database\Repository\RepositoryProvider;
 use App\Database\Repository\RunnerRepository;
-use App\Misc\Util;
+use App\Misc\Util\CommonUtil;
 use Psr\Http\Message\ResponseInterface;
 use Psr\Http\Message\ServerRequestInterface;
 
@@ -61,10 +60,10 @@ class CategoriesResponder implements ResponderInterface
             ];
         }
 
-        Util::insertMetadataInResponseArray($responseData);
-        Util::camelizeApiResponseFields($responseData);
+        CommonUtil::insertMetadataInResponseArray($responseData);
+        CommonUtil::camelizeApiResponseFields($responseData);
 
-        $response->getBody()->write(Util::jsonEncode($responseData));
+        $response->getBody()->write(CommonUtil::jsonEncode($responseData));
 
         return $response;
     }
